Resolved an issue and added tests
If a variable was created with the same name as the synthetic variable an AXSMatcherRef would have been compiled to, such compilation would result in errors.
Now we test the ITreeAppendable contains a name for the reference instead of an object for the given expected name. Also added tests for this